test testCanCalculateTotalValue is passing, lost some time with the if statement, word else has to be used

+++ src/Behaviours/Buyable.java
@@ -2,8 +2,8 @@ package Behaviours;
 
 public interface Buyable {
 
-    int price();
+    float price();
     boolean twoForOne();
-//    int applyTwoForOne();
+
 
 }
